Machine-independent support for split 64-bit floats
This patch introduces an architectural parameter `splitfloat` analogous to
`splitlong`. If true, 64-bit `float` values will be allocated to pairs of
32-bit registers. This is expected to be useful for better register
allocation on ARM as well as for the upcoming MPPA port.

The patch also introduces a `Twofloat` register pair constructor analogous
to `Twolong` and implements the machine-independent parts of the backend
support for this.