MPPA: Function calls
Including somewhat more general memory accesses and the copy instruction.

The MPPA's ABI features a 16-byte scratch space at the bottom of each stack
frame which must be allocated by each non-leaf function and may be used by
its callees (i.e., it is a "caller-saved" stack area). We allocate this area
for ABI compatibility but do not use it ourselves.

Tail calls are not supported yet.