WIP: Represent machine subregisters as parts of registers
In the Locmap representation on LTL and lower levels, introduce "register
parts" to model access to subregister. A register part is a full register or
the upper or lower half of a full register. Full registers are modeled as
holding one or two values, and an access to a subregister part gets/sets the
corresponding half of a pair. Full registers holding one value may only be
read as a whole, and full registers holding a pair may only be read as
subregisters. Other reads are undefined.

FIXME: Rename `mreg_access` to something better like `mreg_part`.
FIXME: Reorganize what goes where; some proofs are very ugly. Use nicer
auxiliary lemmas and tactics everywhere.