POC: Folding of algebraic zeros in value analysis
The function
```
int f(int x, int y) {
    int z = 0;
    return x | (y * z);
}
```
always returns `x`, neither the multiplication nor the bitwise-or is needed.
CompCert removes the multiplication (replacing it by loading 0 into a
register), but leaves the or operation.

The reason for this is that the replacement of `y * 0` by 0 only takes place
at the "strength reduction" stage, which is purely local to the current
operation. Values are propagated before, by the value analysis, but the
value analysis does not contain a rule for folding this expression. Indeed,
it cannot represent this, because the value of `y` may be `Vundef`, and the
value domain's notion of "match" does not allow us to replace an undefined
value by the more defined constant 0.

This patch extends the value domain with a new constructor `Ix` (integer or
don't care, i.e., undefined) and adds just enough machinery to allow us to
handle the above example program correctly. As this involves a nontrivial
amount of code duplication, even just to handle the bitwise-or operator, it
is clearly not the best approach for doing this.