Model the register file as a block of bytes
Introduce a new module `Registerfile` that models the register file as a
block of bytes. Register accesses map register names to addresses in this
block and load/store them using the machinery for memory loads/stores.

The registers' addresses in the register file are computed from their
`IndexedMreg` indices. When computing addresses, these indices are
interpreted as machine words. The indices of 64-bit registers must thus be
spaced at least 2 apart. This patch also adjusts the indices for all
architectures.