DIRTY WIP: Model stack frames as bytes
Similarly to how we model the register file as a kind of memory block that
stores bytes (more precisely, memvals), add a Stack module that does the
same for the stack frame. The three kinds of stack slots (incoming,
outgoing, local) each have their separate memval map that stores the data.

The idea is to allow us to model spilling as copying of bytes from registers
to the stack and back. This is meant to preserve spilled values, even if
those values are in the subregisters of a spilled superregister. This cannot
be modeled as copying values directly, since the value of a superregister is
undefined if one of its subregisters holds a value.

FIXME: Many many admits.