Change types of stack slots to quantities
Replace the `ty` type field in stack slots with a `q` quantity field. The
idea is to make typing more coarse-grained, similarly to registers. This
solves some problems related to copying data between strictly typed slots
and more liberally typed registers.

This patch contains the target-independent parts of this change, as well as
the target-dependent parts for ARM.